requirement.txt
streamlit==0.86.0
Pillow==8.3.2
google==3.0.0
langchain-google-genai
